Is is realy so difficult?

+++ backend/database/models/user.py
@@ -1,9 +1,9 @@
-from sqlalchemy import Column, Integer, LargeBinary
+from sqlalchemy import Column, LargeBinary, BigInteger
 from backend.database.db import Database
 
 
 class User(Database.Base):
     __tablename__ = 'users'
 
-    user_id = Column(Integer, primary_key=True)
+    user_id = Column(BigInteger, primary_key=True)
     mnemonic = Column(LargeBinary, nullable=False)
